test:static: Fix versions for the tools to use
So that an upstream update on a tool (i.e. pylint) does not break the
pipeline.

Let dependabot launch PRs on these too and follow the regular flow.

Setting pylint to a version earlier than latest, as latest version
introduced a new rule that makes our code not to lint.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>